requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.16.1
fsspec==2024.9.0
groq==0.11.0
h11==0.14.0
http-exceptions==0.2.10
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
jiter==0.5.0
numpy==2.1.1
openai==1.47.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tokenizers==0.20.0
toolhouse==1.2.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
